server/requirements.txt
# /server/requirements.txt

fastapi
uvicorn[standard]
SQLAlchemy>=1.4.0
alembic>=1.8.0
aiosqlite>=0.18.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic[email]
pydantic-settings>=2.0.0
email-validator>=2.0.0
httpx>=0.24.0
asyncpg>=0.28.0
redis>=4.6.0
pytest-asyncio>=0.21.0
passlib>=1.7.4
bcrypt>=4.0.1
python-jose>=3.3.0
python-multipart>=0.0.6
aiofiles>=23.0.0
google-generativeai>=0.2.0
sentry-sdk>=1.39.0